chore: organize files and helper functions
* Moved helper functions to src/utils/helpers
* Moved URL constants to src/data/
